src:Redactor: mode crypto in to separate module
Same rationale as the above. A crypto-focused module should allow us to
manage the key generation/management as well as encrypting data on
demand.
